Refactor CLI file
Extract the different CLI functionalities into their own module and
classes. This way, the implementation of the Diagnose command will be
simpler.
